lecture4gachon-main/sample_application/flask/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.1
filelock==3.12.2
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.74
Pillow==10.0.0
requests==2.31.0
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
